// src/main/resources/site/content-types/article/article.ts
// WARNING: This file was automatically generated by "no.item.xp.codegen". You may lose your changes if you edit it.
export interface Article {
  /**
   * Preface
   */
  preface: string;

  /**
   * Body
   */
  body: string;

  /**
   * Image for article
   */
  articleImage?: string;

  /**
   * Alt text for image
   */
  imageAlt?: string;
}